requirements.txt
Django
django-bootstrap-v5
django-allauth
